Issue #75: Move BrowserIWebViewCallbacks into FullscreenCallbacks.
This code only handled fullscreen callbacks so I separated it into a
separate class/file for clarity.
